requirements.txt
alembic==1.10.4
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
email-validator==1.1.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-Validator==1.4.2
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
py-moneyed==0.8.0
pycountry==22.3.5
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2020.1
qrcode==7.4.2
requests==2.30.0
rich==13.4.1
schwifty==2020.9.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==2.0.2
validators==0.20.0
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
